pyproject.toml
[project]
name = "telegram-news-sentiment"
version = "0.1.0"
description = ""
authors = [
    {name = "astessh"}
]
packages = [
    { include = "telegram_news_sentiment" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch (>=2.7.1,<3.0.0)",
    "lightning (>=2.5.1.post0,<3.0.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "fire (>=0.7.0,<0.8.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "mlflow (>=2.22.0,<3.0.0)",
    "dvc (>=3.59.2,<4.0.0)",
    "dvc-gdrive (>=3.0.1,<4.0.0)",
    "cryptography (>=41.0.3)",
    "pyopenssl (>=23.2.0)",
    "gdown (>=5.2.0,<6.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "pytorch-lightning (>=2.5.1.post0,<3.0.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "peft (>=0.10.0)",
    "datasets (>=2.19.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
pre-commit = "^4.2.0"

[tool.black]
line-length = 88
target-version = ["py311"]  # или твоя версия Python

[tool.isort]
profile = "black"
line_length = 88
